Extract undo jump APIs and split tests
